# rincon/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Rincon",
		"description": "",
		"version": "1.0.0",
		"contact": {
			"name": "Bharat Kathi",
			"email": "bharat1031@gmail.com"
		}
	},
	"servers": [
		{
			"url": "http://localhost:10311",
			"description": ""
		}
	],
	"paths": {
		"/rincon/ping/": {
			"get": {
				"summary": "Ping",
				"description": "",
				"operationId": "2188b25c-be41-414d-a1b5-d96784873cbb",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"text/plain": {}
					},
					"required": false
				},
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": []
			}
		},
		"/rincon/services/": {
			"get": {
				"summary": "Get All Services",
				"description": "",
				"operationId": "4f5b78d3-9b8e-41ee-98a8-5d6eaebdcd92",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"text/plain": {}
					},
					"required": false
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer",
												"format": "int64",
												"example": 463334
											},
											"name": {
												"type": "string",
												"example": "rincon"
											},
											"version": {
												"type": "string",
												"example": "0.1.0"
											},
											"endpoint": {
												"type": "string",
												"example": "http://localhost:10311"
											},
											"health_check": {
												"type": "string",
												"example": "http://localhost:10311/rincon/ping"
											},
											"updated_at": {
												"type": "string",
												"example": "2024-05-14T15:18:10.834306-07:00"
											},
											"created_at": {
												"type": "string",
												"example": "2024-05-14T15:18:10.834306-07:00"
											}
										}
									}
								}
							}
						},
						"headers": {
							"Content-Type": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "application/json; charset=utf-8"
								}
							},
							"Content-Length": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "233"
								}
							},
							"Connection": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "close"
								}
							},
							"Date": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "Tue, 14 May 2024 22:18:17 GMT"
								}
							}
						}
					}
				},
				"security": []
			},
			"post": {
				"summary": "Create Service",
				"description": "",
				"operationId": "c52d2034-b67b-4c1a-b6cd-e8042c63af60",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"schema": {
							"type": "string",
							"default": "Basic YWRtaW46YWRtaW4=",
							"description": ""
						}
					},
					{
						"name": "Content-Type",
						"in": "header",
						"schema": {
							"type": "string",
							"default": "application/json; charset=utf-8",
							"description": ""
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"endpoint": {
										"type": "string",
										"example": "localhost:8082"
									},
									"name": {
										"type": "string",
										"example": "Service C"
									},
									"health_check": {
										"type": "string",
										"example": "localhost:8080/health"
									},
									"version": {
										"type": "string",
										"example": "1.2.0"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "integer",
											"format": "int64",
											"example": 680427
										},
										"name": {
											"type": "string",
											"example": "service_c"
										},
										"version": {
											"type": "string",
											"example": "1.2.0"
										},
										"endpoint": {
											"type": "string",
											"example": "localhost:8082"
										},
										"health_check": {
											"type": "string",
											"example": "localhost:8080/health"
										},
										"updated_at": {
											"type": "string",
											"example": "2024-05-14T15:20:18.561061-07:00"
										},
										"created_at": {
											"type": "string",
											"example": "2024-05-14T15:20:18.561061-07:00"
										}
									}
								}
							}
						},
						"headers": {
							"Content-Type": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "application/json; charset=utf-8"
								}
							},
							"Content-Length": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "213"
								}
							},
							"Connection": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "close"
								}
							},
							"Date": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "Tue, 14 May 2024 22:20:18 GMT"
								}
							}
						}
					}
				},
				"security": []
			}
		},
		"/rincon/services/{id}/": {
			"get": {
				"summary": "Get Service",
				"description": "",
				"operationId": "cb1faf22-038e-4746-a9e0-c1f867fd1cb6",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"default": "service_a",
							"description": "ID or Name of registered service"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"text/plain": {}
					},
					"required": false
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "array"
									}
								}
							}
						},
						"headers": {
							"Content-Type": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "application/json; charset=utf-8"
								}
							},
							"Content-Length": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "2"
								}
							},
							"Connection": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "close"
								}
							},
							"Date": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "Tue, 14 May 2024 21:40:22 GMT"
								}
							}
						}
					}
				},
				"security": []
			},
			"delete": {
				"summary": "Remove Service",
				"description": "",
				"operationId": "50a1d0bb-465e-4da3-a014-5aa306639ed1",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"default": "216857",
							"description": "ID of service to remove"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"text/plain": {}
					},
					"required": false
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Service with id 216857 removed"
										}
									}
								}
							}
						},
						"headers": {
							"Content-Type": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "application/json; charset=utf-8"
								}
							},
							"Content-Length": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "44"
								}
							},
							"Connection": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "close"
								}
							},
							"Date": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "Tue, 14 May 2024 22:06:56 GMT"
								}
							}
						}
					}
				},
				"security": []
			}
		},
		"/rincon/routes/": {
			"get": {
				"summary": "Get All Routes",
				"description": "",
				"operationId": "aa83ecba-173c-469e-84ac-e85f5a2f95e4",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"text/plain": {}
					},
					"required": false
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"route": {
												"type": "string",
												"example": "/rincon/ping"
											},
											"service_name": {
												"type": "string",
												"example": "rincon"
											},
											"created_at": {
												"type": "string",
												"example": "2024-05-14T15:00:55.729377-07:00"
											}
										}
									}
								}
							}
						},
						"headers": {
							"Content-Type": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "application/json; charset=utf-8"
								}
							},
							"Content-Length": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "416"
								}
							},
							"Connection": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "close"
								}
							},
							"Date": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "Tue, 14 May 2024 22:02:15 GMT"
								}
							}
						}
					}
				},
				"security": []
			},
			"post": {
				"summary": "Create Route",
				"description": "",
				"operationId": "0b98260b-cc46-4916-aac3-ec317ef17346",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"schema": {
							"type": "string",
							"default": "Basic YWRtaW46YWRtaW4=",
							"description": ""
						}
					},
					{
						"name": "Content-Type",
						"in": "header",
						"schema": {
							"type": "string",
							"default": "application/json; charset=utf-8",
							"description": ""
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"route": {
										"type": "string",
										"example": "/service"
									},
									"service_name": {
										"type": "string",
										"example": "Service A"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"route": {
											"type": "string",
											"example": "/service"
										},
										"service_name": {
											"type": "string",
											"example": "service_a"
										},
										"created_at": {
											"type": "string",
											"example": "2024-05-11T12:48:21.289331-07:00"
										}
									}
								}
							}
						},
						"headers": {
							"Content-Type": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "application/json; charset=utf-8"
								}
							},
							"Content-Length": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "95"
								}
							},
							"Connection": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "close"
								}
							},
							"Date": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "Sat, 11 May 2024 19:48:25 GMT"
								}
							}
						}
					}
				},
				"security": []
			}
		},
		"/rincon/routes/{route}/": {
			"get": {
				"summary": "Get Route",
				"description": "",
				"operationId": "8fb8f557-259e-4ff2-acfd-32bb3e51acd5",
				"parameters": [
					{
						"name": "route",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"default": "rincon<->test",
							"description": ""
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"text/plain": {}
					},
					"required": false
				},
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": []
			}
		},
		"/rincon/services/{id}/routes/": {
			"get": {
				"summary": "Get Routes For Service",
				"description": "",
				"operationId": "7657a466-de5b-481c-9529-b49e7937feb7",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"default": "service_a",
							"description": "ID or Name of registered service"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"text/plain": {}
					},
					"required": false
				},
				"responses": {
					"200": {
						"description": ""
					}
				},
				"security": []
			}
		},
		"/rincon/match/{route}/": {
			"get": {
				"summary": "Match Route",
				"description": "",
				"operationId": "d5931715-4cff-4401-8b0e-5e16fbb41a71",
				"parameters": [
					{
						"name": "route",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"default": "rincon<->test<->wow<->bruh",
							"description": "Endpoint route with \"/\" swapped for \"<->\""
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"text/plain": {}
					},
					"required": false
				},
				"responses": {
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "No route /rincon/test/wow/bruh found"
										}
									}
								}
							}
						},
						"headers": {
							"Content-Type": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "application/json; charset=utf-8"
								}
							},
							"Content-Length": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "50"
								}
							},
							"Connection": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "close"
								}
							},
							"Date": {
								"description": "",
								"schema": {
									"type": "string",
									"example": "Tue, 14 May 2024 22:02:17 GMT"
								}
							}
						}
					}
				},
				"security": []
			}
		}
	},
	"components": {}
}